add Vuong's test (`test_performance_vuong`)
Based on https://github.com/easystats/performance/discussions/183, the Vuong's test seems to be the most useful frequentist alternative as it can accomodate both nested and non-nested models, which makes it quite polyvalent. Unfortunately, it **currently** requires two dependencies, one which is potentially replaceable (sandwich::estfun) and the other one I'm not sure (imhof).

The test is implemented in nonnest2 but only handles paris of models, so I reimplemented using insight and accomodating more than 2 models.