Many updates from pdwaggoner
I have hacked it up a bit, but didn't change any of the substance. Just moved some things around, corrected some grammatical issues, and streamlined some of the language to make it read a bit more fluidly. Of course feel free to revert if you don't like anything, or let me know and we can discuss.