fix: handle multiple colliders in check_dag (#878)
Fix error when check_dag encounters multiple colliders in adjustment
nodes. The issue was that 'collider %in% adjustment_nodes' returns a
vector of logical values when there are multiple colliders, which
cannot be used with the && operator.

Solution: Use any() to properly handle vector comparison.

Also add missing author names to WORDLIST for spell check:
- Dunkler, Graham, Gregorich, Morrissey, Papeş, Pandey, Park,
  Ruxton, Strohmaier

Fixes #878